phase1_backend/requirements_clean.txt
fastapi==0.128.8
uvicorn==0.40.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
python-jose==3.5.0
passlib==1.7.4
bcrypt==3.2.2
python-dotenv==1.2.1
pydantic==2.12.5
email-validator==2.3.0
